Handle B.3.5 function declarations in if statements (#38)
Ugh.

This is for code which looks like

```js
function outer() {
  if (foo) function f(){}
  return f;
}
```

This case is a bit weird. Recap: b33 function declarations normally create two variables, one function-scoped and one block-scoped. Normally we put an assignment of the function to the block-scoped variable at the top of the containing block, and explicate the declaration statement itself as a read from the block-scoped variable and a write to the function-scoped variable.

In this particular case, though, there's no block to hoist to the top of. Instead, compile the declaration to an assignment of the function to the function-scoped variable directly, bypassing the block-scoped variable.